requirements.txt
numpy>=1.21.0
pandas>=1.3.0
nibabel>=3.2.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0
click>=8.0.0
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
